refactor: move all id types to new dedicated module
