+++ honeycomb-core/src/cells/mod.rs
@@ -1,0 +1,13 @@
+//! Module short description
+//!
+//! Should you interact with this module directly?
+//!
+//! Content description if needed
+
+// ------ MODULE DECLARATIONS
+
+pub mod identifiers;
+
+// ------ IMPORTS
+
+// ------ CONTENT
